// src/app/shared/models/snapshot-element.interface.ts
export interface SnapshotElement {
  sid: number;
  name: string;
  topeleven: number;
  whiteskill: number;
  position: number;
}

export interface SnapshotAPIResponseElement {
  id: number;
  playerId: number;
  timestamp: Date;
  user: string;
  sid: number;
  name: string;
  topeleven: number;
  whiteskill: number;
  position: number;
}